\name{tstpkg-package}
\alias{tstpkg-package}
\alias{tstpkg}
\docType{package}
\title{\packageTitle{tstpkg}}
\description{\packageDescription{tstpkg}}
\section{Package Content}{\packageIndices{tstpkg}}
\author{\packageAuthor{tstpkg}}
\section{Maintainer}{\packageMaintainer{tstpkg}}
\keyword{package}
